Use the last selection character for Visual mode paste position checks
Visual mode paste position checks must be done for the last character of
the Visual selection. Indeed, the selection may be 'backwards', e.g.
with the `vB` or `Vk` command. In this case position checks for the
cursor will not take into account that the selection can possibly span
over the last column/line.