Avoid auto-formatting in blockwise Visual mode single line pastes
The current paste command is `"_c\<C-R> . EasyClip#GetDefaultReg()`. It
can yield undesired results under certain conditions. For example, tabs
in yanked text are replaced with spaces with `expandtab, ts=4, sts=0`.
This happens because `CTRL-R` in Insert mode inserts contents of a
register with respect to various Vim formatting options and we are
brought to Insert mode by the `"_c` command.

As stated in the help for i_CTRL-R:
```
If you have options like 'textwidth', 'formatoptions', or
'autoindent' set, this will influence what will be inserted.
```

Since most people probably want their text pasted unspoiled the proper
paste command is `"_c\<C-R>\<C-O>" . EasyClip#GetDefaultReg()`.

As stated in the help for i_CTRL-R_CTRL-O:
```
Insert the contents of a register literally and don't auto-indent.
```

This commit changes the paste command as described above.