Proper blockwise Visual mode paste
In blockwise Visual mode only the first line of the selection is
important [0], [1]. Fix the code to analyze the first line of the
selection, not the last one

Vim doesn't guarantee that "'<" mark always points to the first
character of the selection. Similarly, "'>" doesn't always point
to the last character of the selection. In blockwise mode these
marks can swap each other's places. This is why max() and min()
are needed.

[0]: https://github.com/svermeulen/vim-easyclip/wiki/Details-of-Visual-mode-paste#blockwise-visual-mode
[1]: https://github.com/svermeulen/vim-easyclip/wiki/Details-of-Visual-mode-paste#blockwise-visual-mode-1