Another fix for characterwise Visual mode paste
If the selection ends at the EOL character of a line and the next line
is empty (`^$`), then after deletion the cursor is placed on the
previous character before the first character of the selection. Thus in
this case we should do pasting after, not before.

Note that if `lnum` points to the last line, then `col([lnum + 1, '$'])`
returns `0`. So there is no need for an additional check if we are on
the last line.

See also: https://github.com/svermeulen/vim-easyclip/wiki/Details-of-Visual-mode-paste#characterwise-visual-mode